update: Draft on how sticky would look for header and banner
