+++ client/components/jetpack/intro-pricing-banner/style.scss
@@ -81,7 +81,7 @@
 
 .is-section-jetpack-connect .intro-pricing-banner.is-sticky,
 .is-group-jetpack-cloud.is-section-jetpack-cloud-pricing .intro-pricing-banner.is-sticky {
-	top: 0;
+	top: 46px;
 }
 
 // target Calypso pricing filter bar, not jetpack cloud
