+++ client/jetpack-cloud/sections/pricing/jpcom-masterbar/cloud-cart/cloud-cart-icon.tsx
